chore: extract types, state, and utils from spell grant components
Extract interfaces to types/components/ files, reactive state to
co-located .svelte.ts files, and use destructuring for functions
and $derived for reactive getters. Remove duplicate
LevelUpSchoolSelection in favor of shared SchoolSelectionGroup.
